Be consistent about line/linestring usage. Move it all to Enum-land instead.
