+++ tilequeue/query/__init__.py
@@ -141,7 +141,7 @@ def _make_rawr_fetcher(cfg, layer_data, query_cfg, io_pool):
 
 
 def _make_layer_info(layer_data, process_yaml_cfg):
-    from tilequeue.query.common import LayerInfo
+    from tilequeue.query.common import LayerInfo, parse_shape_types
 
     layers = {}
     functions = _parse_yaml_functions(process_yaml_cfg)
@@ -149,21 +149,11 @@ def _make_layer_info(layer_data, process_yaml_cfg):
     for layer_datum in layer_data:
         name = layer_datum['name']
         min_zoom_fn, props_fn = functions[name]
-        shape_types = _parse_shape_types(layer_datum['geometry_types'])
+        shape_types = parse_shape_types(layer_datum['geometry_types'])
         layer_info = LayerInfo(min_zoom_fn, props_fn, shape_types)
         layers[name] = layer_info
 
     return layers
-
-
-def _parse_shape_types(inputs):
-    outputs = set()
-    for value in inputs:
-        if value.startswith('Multi'):
-            value = value[len('Multi'):]
-        outputs.add(value.lower())
-
-    return outputs or None
 
 
 def _parse_yaml_functions(process_yaml_cfg):
